+++ src/app/modules/search/components/data-filter/data-filter.component.spec.ts
@@ -1,0 +1,50 @@
+/**
+ * Copyright 2021 Vitagroup AG
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+import { ComponentFixture, TestBed } from '@angular/core/testing'
+import { TranslateModule } from '@ngx-translate/core'
+import { DataFilterService } from 'src/app/core/services/data-filter/data-filter.service'
+
+import { DataFilterComponent } from './data-filter.component'
+
+describe('DataFilterComponent', () => {
+  let component: DataFilterComponent
+  let fixture: ComponentFixture<DataFilterComponent>
+
+  const mockDataFilterService: DataFilterService = {}
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [DataFilterComponent],
+      imports: [TranslateModule.forRoot()],
+      providers: [
+        {
+          provide: DataFilterService,
+          useValue: mockDataFilterService,
+        },
+      ],
+    }).compileComponents()
+  })
+
+  beforeEach(() => {
+    fixture = TestBed.createComponent(DataFilterComponent)
+    component = fixture.componentInstance
+    fixture.detectChanges()
+  })
+
+  it('should create', () => {
+    expect(component).toBeTruthy()
+  })
+})