+++ CHANGELOG.md
@@ -15,6 +15,7 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 - AQL Category select in AQL editor dialog ([#229])
 - Cohort-Builder disabled state ([#234])
 - Cohort-Builder and accordion style to data-explorer ([#235])
+- Search-Page to filter patients for possible projects ([#238])
 
 ### Changes
 
@@ -466,3 +467,4 @@ The format is based on [Keep a Changelog][keep a changelog] and this project adh
 [#233]: https://github.com/NUM-Forschungsdatenplattform/num-portal-webapp/pull/233
 [#234]: https://github.com/NUM-Forschungsdatenplattform/num-portal-webapp/pull/234
 [#235]: https://github.com/NUM-Forschungsdatenplattform/num-portal-webapp/pull/235
+[#238]: https://github.com/NUM-Forschungsdatenplattform/num-portal-webapp/pull/238